+++ actix-web-lab/src/cbor.rs
@@ -2,7 +2,7 @@
 
 use actix_web::{HttpRequest, HttpResponse, Responder};
 use bytes::Bytes;
-use derive_more::{Deref, DerefMut, Display};
+use derive_more::Display;
 use mime::Mime;
 use once_cell::sync::Lazy;
 use serde::Serialize;
@@ -10,8 +10,10 @@ use serde::Serialize;
 static CBOR_MIME: Lazy<Mime> = Lazy::new(|| "application/cbor".parse().unwrap());
 
 /// CBOR responder.
-#[derive(Debug, Deref, DerefMut, Display)]
+#[derive(Debug, Display)]
 pub struct Cbor<T>(pub T);
+
+impl_more::impl_deref_and_mut!(<T> in Cbor<T> => T);
 
 impl<T: Serialize> Responder for Cbor<T> {
     type Body = Bytes;